refactor(api): swagger 어노테이션을 프로덕션 코드에서 제거하기 위해 상속을 이용한다.
