+++ resumarble-api/src/main/kotlin/resumarble/api/prediction/PredictionController.kt
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import resumarble.api.global.response.Response
+import resumarble.api.swagger.SwaggerPredictionWebPort
 import resumarble.core.domain.prediction.application.port.`in`.FindPredictionUseCase
 import resumarble.core.domain.prediction.application.port.`in`.PredictionResponse
 
@@ -12,10 +13,10 @@ import resumarble.core.domain.prediction.application.port.`in`.PredictionRespons
 @RequestMapping("/predictions")
 class PredictionController(
     private val findPredictionUseCase: FindPredictionUseCase
-) {
+) : SwaggerPredictionWebPort {
 
     @GetMapping("/{userId}")
-    fun getPredictionByUserId(@PathVariable userId: Long): Response<List<PredictionResponse>> {
+    override fun getPredictionByUserId(@PathVariable userId: Long): Response<List<PredictionResponse>> {
         return Response.ok(findPredictionUseCase.getPredictionByUserId(userId))
     }
 }
